fixed Example 2 - disable Intune portion of EMS
Simplified the commands: removed the requirement to remove the license from already licensed user, and removed the -addLicense parameter from Set-MsolUserLicese (it is not required if you are goint to change the license portion only, the license itself is inherited from the MsolLicenseOptions object). The -addLicense need to be kept only if a license is not already assigned.