Update the Note in Step 1
Update the Note in Step 1 to clarify and expand the definition of what exactly "VPN Profile name" means wrt the actual field used during the configuration of the VPN profile. This will prevent support calls when customers use the wrong name when configuring the OMA-URI below.

This is specific to the Azure portal. A separate pull request for updating the same Note for the Classic Portal has been opened here: https://github.com/Microsoft/IntuneDocs/pull/472