Re-adding 4 settings that were added to Intune in May
The min os, min app settings were under the docs for Android app protection policies around May when the feature went live. But if I go in the docs today, I only see them on iOS. Not sure what happened here, pinged iX about it at the start of September. These are features that are live and available to customers, and as such I added them back.