M16655: Possible incorrect string: Download the Install and sign in to the Intune Company Portal app
Hello, @barlanmsft
Translator has reported possible source content issue. It seems it should say "Download and install Intune Company Portal app." "sign in"  should be omitted because it appears in the next step.

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 
Many thanks in advance.