Removed split tunnel from the list of AfW VPN settings
The split tunnel option does not appear in the Intune Azure portal; we don't support this option on the back-end, either, and my understanding is that app config is the only way to do this and it only works for some VPN clients. Having the split tunnel option in the UI is incorrect and is confusing customers.