Updated platform support statement
I know these are going to be retired soon, but customers keep bugging us about having this info out of date so I changed this to align with the info in the Azure docs - the info also applies to the classic portal. Didn't update date in case other tidbits in this doc are out-of-date.