Added Cisco AnyConnect and minor cleanup
With the release of support for the newer Cisco AnyConnect client, we need to add Cisco AnyConnect to the official per-app VPN support list. I not only added Cisco AnyConnect, I changed the order of the list to alphabetical so as to not show preference, and I made another small grammar edit.