M27205: Splitting paragraph from last list item
Hello, @ErikjeMS ,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
"This segment should probably be split after "not-applicable". The last item in the list of states is immediately followed by a sentence that is not part of the list."
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.

Many thanks in advance.