Updated example scenario for DEM
The old wording implies that a DEM account is created and the supervisor somehow 'added to' that account, giving the supervisor DEM rights.  The way it actually works is a new DEM account is created, and the supervisor can use that account to enroll devices.  It's not like adding an existing user to a group to grant the user DEM permissions, it's a second account for that user which is used only for DEM purposes.