broken link when published to docs.microsoft.com
when viewed via docs.microsoft.com domain the link was broken due the .md at the end. Published it looked like this:
https://docs.microsoft.com/en-us/intune-user-help/enroll-your-device-in-intune-android.md
but must be this:
https://docs.microsoft.com/en-us/intune-user-help/enroll-your-device-in-intune-android
I think because of the prefix /intune-user-help/ the .md link is not correctly replaced.
/intune-user-help/enroll-your-device-in-intune-android.md
therefore changing it to 
/intune-user-help/enroll-your-device-in-intune-android
this should work on the published site.