M29418: Possible typo Open-In/Open in
Hello, @Erikre,
Translator has reported possible source content issue. 
"Linguist comment: Could you please confirm if the action extensions are of the type "Open-In" (we used this solution) or if there is simply a typo and it should say "Yes to open in action extensions"?"

Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.