Make configuring NSFaceIDUsageDescription Info.plist key an explicit integration step
Making configuration of the NSFaceIDUsageDescription Info.plist key an explicit step, so it's harder to miss. This also keeps our Xamarin bindings documentation accurate as it references these instructions, but tells developers to skip the final step (back when it only included info on running IntuneMAMConfigurator). Xamarin developers will also need to configure this key.