Note for known issue for per-app VPN w/F5
F5 Access for Android 3.0.3 has known issues that prevent per-app VPN from working properly in certain scenarios. F5 has published these in the release notes for the app; we should link to that doc from our own doc to acknowledge that we're aware of these issues and for convenience in providing the customer with more info.