C44395: Asterisk issue
Hello, @Erikre,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
Description of the source issue: there are extra spaces as well as plain asterisks. It is not clear whether it was meant to be a footnote symbol or a mistyping for italic type.
Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.