Update intune-management-extension.md
Changed introduction paragraph "Move to modern management" as it had some advices to install Win32 with PowerShell because of missing features. This is not true anymore. Win32 app support is there. The Last Note about assigning to computer groups not possible and then aad device groups is possible seems to be a mistake it is a contradiction. resolve it as it is possible to target aad device groups now.