Update device-restrictions-ios.md
Adding clarification for functionality that show or hide apps settings supports, but it was confusing for customers who were not sure which types of apps you can show/hide (LOB, built-in, store apps). We can show.hide all three types, and that is what Apple supports as well. To hide/show a built-in or LOB app, the it admin must specify the bundle ID and app name of the app.