Update app-configuration-policies-managed-app.md
We're hearing from customers that the prerequisite of having app protection policies targeted to an app before app config policies that effect isn't clear. And we're hearing that this is the doc most customers work off of when setting up their app config policies. Borrowing some text from the Managed Browser/Edge docs to make this clear.