Ambiguous language surrounding Windows 10 version
"Windows 10 with the October 2018 update." could be read as having the "October 2018" cumulative update rather than the October 2018 FEATURE update. The link did clarify, but there's no reason why this shouldn't be called out directly in the document. Now that v1903 is being released as well, I've added "or greater" to reflect multiple version support.