Add app config note for Palo Alto/Citrix
Palo Alto has been asking that we officially note in our docs that app configuration policy can be used to configure their client, since the "No" in the table is confusing their customers and support staff. Since this also applies to Citrix, I've made a note that can be used for both Palo Alto and Citrix and referenced back to it in the table. If there's a more elegant way to do this, feel free to edit what I have here and/or let me know!