Update endpoint-protection-macos.md
For macOS Catalina and up (10.15+), FileVault only works on devices that have user-approved enrollment, meaning the end user had to physically approve enrollment on their device. These settings won't worn on 10.15+ devices if the user did not manually approve of the MDM enrollment. This should be added as a note at the top because it is important.