GHI 2662: "Setting compliance report" section note
https://github.com/MicrosoftDocs/IntuneDocs/issues/2662

Removed the following:

Windows 10 devices that are Azure AD joined may show the System Account as a non-compliant user. This is expected behavior and doesn't affect the overall device compliance. 

Replaced it with the following:

A policy can be assigned to a device, and a user on that same device. In some scenarios, a device may sync before the user signs in, such as when the device reboots. Compliance may evaluate this user, and show the device as non compliant. This behavior may also show the System Account as a non-compliant user.

This is a known issue with multi-user Windows 10 devices. Any updates are announced in [in development](in-development.md) and/or [what's new](whats-new.md).