Update endpoint-protection-macos.md
Tips for links:   Our doc validation has a strict'ish requirement on  'more information' references. They should start as "For more information, see...."   So I edited that.  For the link, if we can use a friendly name we should, and then for URLs, if its in our intune docs repository, we can just use the file name (with .md) without the full path.  So I changed the link to use that friendly format.

One other tip is we should never include the language path of a URL, in this case it was en-us, unless that is an external link (non-microsoft).  When we leave out the language path to our own docs, the page will render in the language of the users browser, and then default back to English if we don't localize content in that language.   IF we set a language path, in a URL, we are forcing the page to render in that language, even if they were viewing the content in a different language.  so we leave it out to enable them to have a better viewing experience.