Update encryption-monitor.md
With the 1909 release of the web CP, the get recovery key text is at the bottom of the devices page for the mac, just like for ipads. So it is no longer necessary to click on the 3 dots first. This will be with the release of 1909 so could we release this doc change in mid september?