Update device-restrictions-ios.md
For user enrollment which is being released today - Devices with a passcode policy greater than 6 digits that are user enrolled will receive a 6-digit requirement, if previously the passcode was set to greater than 6. Devices with a complex passcode requirement that are user enrolled will also receive a 6-digit requirement if previously set to greater than 6. All other passcode policy requirements will not be enforced on user enrolled devices.