Update software-updates-ios.md
I had a call with a customer that said this part is confusing. It is good to point out that software update policies override any device restrictions. So if you have both set, delayed visibility of updates as a restriction and a sw update policy, the policy will win every time.