Update device-restrictions-ios.md
The only settings that apple to user enrolled devices are : Require a passcode, block simple passcode, and minimum of digits. The categories are wrong in the console so we need to point this out. Also, the maximim digits thats enforced on a user enrolled device is 6, but the user can make it more on their own but it can't be enforced by intune to be greater than 6.